rag-search-project/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
numpy==1.26.4
passlib[bcrypt]
python-multipart
pandas
ftfy
PyMuPDF
tqdm
tiktoken
python-docx
pytesseract
pdf2image
Pillow
opencv-python-headless
httpx
readability-lxml
pytest
pytest-asyncio
aiofiles
pydantic
openai
matplotlib
seaborn
PyJWT
bcrypt
slowapi
